+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * removed deprecated alias `entity_base()`. use `declarative_base()` instead.
     * fix imports.
 + Module `functools`: add `cooldown()`
++ Module `color`: add `ColorFormatter()`
 
 ## 0.12.6